Add deadline handler in the Console class
This method handles the ID and Date parsing when adding a deadline to the specified Task. If either the ID or the Date are invalid, an exception is thrown. Otherwise, the deadline is added to a task, given it exists, otherwise a message informing the user that the task with the given ID is non-existent.