Add viewByDeadline to Console class
This method is used to display all tasks grouped based on their deadline and project, respectively. The dates will be sorted in ascending order, while the project names are sorted lexicographically. Tasks with no deadlines will be displayed at the end -- if any such exist -- and will also be grouped based on their project name.