Add getTasksByDeadline to TaskList
This method returns a Map containing Dates as keys in an ascending order. The values a Map of all tasks associated with that deadline date, and their project names. This method is used when executing the "view-by-deadline" command, but only returns the first part of the output, i.e. tasks with deadlines.